Move code_mapping file
