+++ src/sentry/api/endpoints/project_repo_path_parsing.py
@@ -12,7 +12,7 @@ from sentry.api.serializers.rest_framework.base import CamelSnakeSerializer
 from sentry.integrations.base import IntegrationFeatures
 from sentry.integrations.manager import default_manager as integrations
 from sentry.integrations.services.integration import RpcIntegration, integration_service
-from sentry.integrations.utils.code_mapping import find_roots
+from sentry.issues.auto_source_code_config.code_mapping import find_roots
 from sentry.models.repository import Repository
